requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
blinker==1.9.0
whitenoise
CacheControl==0.14.2
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
daphne==4.1.2
dj-database-url==3.0.0
Django==5.1.3
django-jazzmin==3.0.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
docutils==0.21.2
emoji==2.14.0
et_xmlfile==2.0.0
firebase-admin==6.6.0
fonttools==4.55.0
frozenlist==1.5.0
gcloud==0.18.3
geographiclib==2.0
geopy==2.4.1
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
grep==0.3.2
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
jwcrypto==1.5.6
kiwisolver==1.4.7
matplotlib==3.9.2
msgpack==1.1.0
multidict==6.1.0
mysql==0.0.3
mysql-connector-python==9.2.0
mysqlclient==2.2.6
numpy==2.1.3
oauth2client==4.1.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.18.0
PyJWT==2.10.1
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pyparsing==3.2.0
Pyrebase4==4.8.0
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-jwt==4.1.0
pytz==2024.2
redis==5.2.1
requests==2.32.3
requests-toolbelt==0.10.1
robo==0.5.6
robot==20071211
rsa==4.9
selenium==4.27.1
service-identity==24.2.0
setuptools==75.8.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.14.2
sqlparse==0.5.2
trio==0.27.0
trio-websocket==0.11.1
twilio==9.4.6
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20
websocket-client==1.8.0
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.18.3
zope.interface==7.2
gunicorn==21.2.0
python-dotenv
